backend/requirements.txt
Flask==3.0.3
mongoengine==0.28.2
pymongo==3.12.3
Flask-Cors==4.0.1
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
twilio
geopy
dnspython==2.6.1
certifi==2024.7.4
Werkzeug==3.0.3
Flask-Mail==0.10.0 